Refactor and enhance BLLs, controllers, and queries
- Updated `using` statements to include new dependencies.
- Changed `idOperador` data type from `int` to `long` for scalability.
- Improved exception handling with `ValidacaoException`.
- Refactored SQL queries and LINQ for better filtering and readability.
- Enhanced prioritization logic in `ChamadaBLL.cs`.
- Added `evento` parameter to `SetHistorico` in `OperadorHistoricoBLL.cs`.
- Updated `siagDB` connection string in `appsettings.json`.
- Simplified controller responses and error handling.
- Removed redundant code and improved formatting for clarity.

+++ SIAG-CRATO/BLLs/Equipamento/EquipamentoBLL.cs
@@ -133,7 +133,7 @@ public class EquipamentoBLL
         return id > 0;
     }
 
-    public static async Task<bool> SetEquipamentoOperador(int id_operador, int id_equipamento)
+    public static async Task<bool> SetEquipamentoOperador(long id_operador, int id_equipamento)
     {
         using var conexao = new SqlConnection(Global.Conexao);
         var result = await conexao.ExecuteAsync(EquipamentoQuery.UPDATE_EQUIPAMENTO_OPERADOR, new { id_operador, id_equipamento });
