+++ SIAG-CRATO/DTOs/Operador/OperadorLoginDTO.cs
@@ -2,6 +2,6 @@
 
 public class OperadorLoginDTO
 {
-    public int IdOperador { get; set; }
+    public long IdOperador { get; set; }
     public int IdEquipamento { get; set; }
 }
